+++ src/mainviews/detail-view/job-definition.tsx
@@ -13,7 +13,7 @@ import { ListJobsTable } from '../list-jobs';
 import { Scheduler as SchedulerTokens } from '../../tokens';
 
 import { Button, Card, CardContent, FormLabel, Stack } from '@mui/material';
-import { ConfirmDeleteButton } from '../../components/confirm-buttons';
+import { ConfirmDialogDeleteButton } from '../../components/confirm-dialog-buttons';
 import { JupyterFrontEnd } from '@jupyterlab/application';
 import {
   ILabeledValueProps,
@@ -97,7 +97,7 @@ export function JobDefinition(props: IJobDefinitionProps): JSX.Element {
       >
         {trans.__('Edit Job Definition')}
       </Button>
-      <ConfirmDeleteButton
+      <ConfirmDialogDeleteButton
         handleDelete={handleDeleteJobDefinition}
         title={trans.__('Delete Job Definition')}
         dialogText={trans.__(
