+++ src/mainviews/detail-view/job-detail.tsx
@@ -3,9 +3,9 @@ import React, { useState } from 'react';
 import { JupyterFrontEnd } from '@jupyterlab/application';
 
 import {
-  ConfirmDeleteButton,
-  ConfirmStopButton
-} from '../../components/confirm-buttons';
+  ConfirmDialogDeleteButton,
+  ConfirmDialogStopButton
+} from '../../components/confirm-dialog-buttons';
 import { JobFileLink } from '../../components/job-file-link';
 import { Scheduler, SchedulerService } from '../../handler';
 import { useTranslator } from '../../hooks';
@@ -100,13 +100,13 @@ export function JobDetail(props: IJobDetailProps): JSX.Element {
         </Button>
       )}
       {props.model.status === 'IN_PROGRESS' && (
-        <ConfirmStopButton
+        <ConfirmDialogStopButton
           handleStop={handleStopJob}
           title={trans.__('Stop Job')}
           dialogText={trans.__('Are you sure that you want to stop this job?')}
         />
       )}
-      <ConfirmDeleteButton
+      <ConfirmDialogDeleteButton
         handleDelete={handleDeleteJob}
         title={trans.__('Delete Job')}
         dialogText={trans.__('Are you sure that you want to delete this job?')}
